pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "riptide-engine-docker"
version = "0.10.0"
description = "Tool to manage development environments for web applications using containers - Docker Implementation"
readme = "README.rst"
requires-python = ">=3.11"
license = { file = "LICENSE" }
authors = [
    { name = "Marco Köpcke", email = "hello@capypara.de" }
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]
dependencies = [
    "riptide-lib >= 0.10, < 0.11",
    "docker >= 7.1"
]

[project.urls]
Repository = "https://github.com/theCapypara/riptide-engine-docker"
Documentation = "https://riptide-docs.readthedocs.io"

[project.entry-points."riptide.engine"]
docker = "riptide_engine_docker.engine:DockerEngine"

[project.entry-points."riptide.engine.tests"]
docker = "riptide_engine_docker.tests.integration.tester:DockerEngineTester"

[tool.ruff]
line-length = 120

[tools.ruff.lint]
select = ["W", "E", "F", "ARG"]

[tool.mypy]
warn_unused_configs = true
mypy_path = "."
explicit_package_bases = true
namespace_packages = true
check_untyped_defs = true
